deep-research/requirements.txt
openai>=1.0.0
selenium>=4.0.0
beautifulsoup4>=4.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.0.0
markdownify>=0.11.0
tenacity>=8.0.0
rich>=10.0.0 